stackoverflow_clone/users/forms.py

from django.contrib.auth.forms import UserCreationForm, AuthenticationForm
from django.contrib.auth.models import User


class RegisterUserForm(UserCreationForm):

    error_css_class = "form_error"

    def __init__(self, *args, **kwargs):
        super().__init__(*args, **kwargs)
        for field in self.fields.values():
            field.widget.attrs.update([
                ["placeholder", field.label],
                ["class" ,"register_input_widget"]
            ])

    class Meta:
        model = User
        fields = ['username', 'password1', 'password2']


class LoginUserForm(AuthenticationForm):

    error_css_class = "form_error"

    def __init__(self, *args, **kwargs):
        super().__init__(*args, **kwargs)
        for field in self.fields.values():
            field.widget.attrs.update([
                ["placeholder", field.label],
                ["class", "login_input_widget"]
            ])
